rio: new clock type, time functions including rfc 3339 parser
New rio_clock_time_t struct represents time as exactly as the conventional
nanosecond uint128_t in a more parse and comparison friendly structure. It is
not a perfect abstraction. We probably want to go towards orthonology with
the rio_date_time_t struct, which covers similar use cases.

There are also functions to convert to and from rio_clock_time_t and some new
helpers.