switch len and str position in fstr_t
This might seem like a trivial change that is not needed but it has several
advantages. First of all it makes a fstr_t binary compatible with the Linux
struct iovec. Secondly it should create more optimized code when following the
string pointer.

Lastly, it forces a compile error when attempting to define a fstr_t from an
integer as it can no longer be falsely interpreted as the initializer for the
first field (len) in the struct (since an integer cannot implicitly be cast to
a pointer).

E.g. this no longer compiles:

struct {fstr_t x;} foo[] = {{"x"}, {"y"}, {3}};